Don't break backwards compatibility
