+++ bls-demo.html
@@ -5,7 +5,7 @@
 <title>BLS signature sample</title>
 <link rel="stylesheet" href="style.css" type="text/css">
 <script type='text/javascript' src="./bls_c.js"></script>
-<script type='text/javascript' src="./bls.js"></script>
+<script type='text/javascript' src="./bls-factory.js"></script>
 <script type='text/javascript' src="./bls-demo.js"></script>
 </script>
 </head>
